Changed some funcs and started adding docs
Changed a lot of functions in database.rs to take references as these
funcs do not need to consume the data they are given. Also started
adding doc to functions outside of those neccessary for structopt.
Changed operation such that operations that read data from the db return
a result and operations that don't read data from the db just return a
value.